Added Buffer Enemy Behavior (buggy, will fix)
+Added buffer enemy behavior
*Buffer enemies have a maxBuffDistance and a list of enemies to buff
called a "squad"

+++ SGP_NightBrawler_Strawman/Assets/Resources/Scripts/Modifier/MOD_Base.cs
@@ -4,14 +4,14 @@ using System.Collections.Generic;
 
 public class MOD_Base : Item {
 
-    public bool Mod_IsBuff;             //Is true if its a Buff and False if is a Debuff
-    public bool Mod_PartyWide;          //Dose it affect the whole party or only the current player.
+    public bool Mod_IsBuff;             //Is true if it's a Buff and False if it's a Debuff
+    public bool Mod_PartyWide;          //Does it affect the whole party? Or only the current player.
 
     public int Mod_ModIndexNum;         //The Number ID of the Effect 
     public float Mod_effectTimer;       //How long the effect will last
 
     public PlayerController Mod_Actor;  //The Actor been aflicted with the Effect
-    
+	public ACT_Enemy enemy;
     
     // Use this for initialization
     void Start()
@@ -36,21 +36,43 @@ public class MOD_Base : Item {
 	// Update is called once per frame
     public override void Update()
     {
-        if (Mod_Actor != null)                      //If No Actor is selected just ignore the update
-        {
-            ModifyActor();                          //Actually affect the Actor
-            Mod_effectTimer -= Time.deltaTime;      //Reduce the Time of the effect
-            if (Mod_effectTimer <= 0.0f)
-                EndModifyActor();                   //Once Timer is over kill the effect.
-        }
-        else
-            Mod_Actor = transform.gameObject.GetComponent<PlayerController>();      //Get the Actor once attach to the effect
+		bool isEnemy = false;
+		if (enemy != null)
+		{
+			isEnemy = true;
+			ModifyEnemy();
+			Mod_effectTimer -= Time.deltaTime;
+			if (Mod_effectTimer <= 0.0f)
+				EndModifyEnemy();
+		}
+		else
+		{
+			enemy = transform.gameObject.GetComponent<ACT_Enemy>();
+			isEnemy = true;
+		}
+		if (!isEnemy)
+		{
+			if (Mod_Actor != null)                      //If No Actor is selected just ignore the update
+			{
+				ModifyActor();                          //Actually affect the Actor
+				Mod_effectTimer -= Time.deltaTime;      //Reduce the Time of the effect
+				if (Mod_effectTimer <= 0.0f)
+					EndModifyActor();                   //Once Timer is over kill the effect.
+			}
+			else
+				Mod_Actor = transform.gameObject.GetComponent<PlayerController>();  //Get the Actor once attach to the effect
+		}    
 	}
 
     public virtual void ModifyActor()   //Just a virtual fuction for its children
     {
         
-    }   
+    }
+
+	public virtual void ModifyEnemy()
+	{
+		
+	}
 
     public virtual void EndModifyActor()    //Reset the characte's HasMod Veriables.
     {
@@ -61,7 +83,12 @@ public class MOD_Base : Item {
                 Mod_Actor.party[i].Act_HasMod = false;
     }
 
-    public void SetModEffect(bool _IsItBuff, int _IndexNum)     //Selects the Buff or Debuff from the list
+	public virtual void EndModifyEnemy()    //Reset the characte's HasMod Veriables.
+	{
+		enemy.Act_HasMod = false;
+	}
+
+    public void SetModEffectPlayer(bool _IsItBuff, int _IndexNum)     //Selects the Buff or Debuff from the list
     {
         if (_IsItBuff)
         {
@@ -110,7 +137,56 @@ public class MOD_Base : Item {
         }
     }
 
-    public bool NullNewEffects()        //Checks if the current actor already has an effect 
+	public void SetModEffectEnemy(bool _IsItBuff, int _IndexNum)     //Selects the Buff or Debuff from the list
+	{
+		if (_IsItBuff)
+		{
+			switch (_IndexNum)
+			{
+				case 0:
+					enemy.gameObject.AddComponent<MOD_CDDecrease>();
+					break;
+				case 1:
+					enemy.gameObject.AddComponent<MOD_DMGIncrease>();
+					break;
+				case 2:
+					enemy.gameObject.AddComponent<MOD_DMGProtection>();
+					break;
+				case 3:
+					enemy.gameObject.AddComponent<MOD_HPInstant>();
+					break;
+				case 4:
+					enemy.gameObject.AddComponent<MOD_HPRegen>();
+					break;
+				case 5:
+					enemy.gameObject.AddComponent<MOD_SPDIncrease>();
+					break;
+			}
+		}
+		else if (!_IsItBuff)
+		{
+			switch (_IndexNum)
+			{
+				case 0:
+					enemy.gameObject.AddComponent<MOD_DMGDecrease>();
+					break;
+				case 1:
+					enemy.gameObject.AddComponent<MOD_DMGIncomingIncrease>();
+					break;
+				case 2:
+					enemy.gameObject.AddComponent<MOD_DoT>();
+					break;
+				case 3:
+					enemy.gameObject.AddComponent<MOD_Slowed>();
+					break;
+				case 4:
+					enemy.gameObject.AddComponent<MOD_Stunned>();
+					break;
+			}
+		}
+	}
+
+    public bool NullNewEffectsPlayer()        //Checks if the current actor already has an effect 
     {
         if (Mod_Actor.party[Mod_Actor.currChar].Act_HasMod)
         {
@@ -132,4 +208,23 @@ public class MOD_Base : Item {
         }
         
     }
+
+	public bool NullNewEffectsEnemy()        //Checks if the current actor already has an effect 
+	{
+		if (enemy.Act_HasMod)
+		{
+			if (enemy.Act_ModIsBuff != Mod_IsBuff)
+			{
+				//Remove All effects in Character.
+				Debug.Log("Get Reck Potion");
+			}
+			return true;
+		}
+		else
+		{
+			enemy.Act_HasMod = true;
+			return false;
+		}
+
+	}
 }
